Crate hipvs

cuVS: Rust bindings for Vector Search on the GPU

This crate provides Rust bindings for cuVS, allowing you to run approximate nearest neighbors search on the GPU.

Modules

brute_force

Brute Force KNN

cagra

CAGRA is a graph-based nearest neighbors implementation with state-of-the art query performance for both small- and large-batch sized search.

cluster

distance

distance_type

ivf_flat

The IVF-Flat method is an ANN algorithm. It uses an inverted file index (IVF) with unmodified (that is, flat) vectors. This algorithm provides simple knobs to reduce the overall search space and to trade-off accuracy for speed.

ivf_pq

Inverted File Product Quantization

Structs

ManagedTensor

ManagedTensor is a wrapper around a dlpack DLManagedTensor object. This lets you pass matrices in device or host memory into cuvs.

Resources

Resources are objects that are shared between function calls, and includes things like CUDA streams, cuBLAS handles and other resources that are expensive to create.

Enums

Error

Type Aliases

Result