hipvs::distance_type

Type Alias DistanceType

pub type DistanceType = cuvsDistanceType;

Aliased Type

#[repr(u32)]
pub enum DistanceType {

    L2Expanded = 0,
    L2SqrtExpanded = 1,
    CosineExpanded = 2,
    L1 = 3,
    L2Unexpanded = 4,
    L2SqrtUnexpanded = 5,
    InnerProduct = 6,
    Linf = 7,
    Canberra = 8,
    LpUnexpanded = 9,
    CorrelationExpanded = 10,
    JaccardExpanded = 11,
    HellingerExpanded = 12,
    Haversine = 13,
    BrayCurtis = 14,
    JensenShannon = 15,
    HammingUnexpanded = 16,
    KLDivergence = 17,
    RusselRaoExpanded = 18,
    DiceExpanded = 19,
    BitwiseHamming = 20,
    Precomputed = 100,

}

Variants

L2Expanded = 0

evaluate as dist_ij = sum(x_ik^2) + sum(y_ij)^2 - 2*sum(x_ik * y_jk)

L2SqrtExpanded = 1

same as above, but inside the epilogue, perform square root operation

CosineExpanded = 2

cosine distance

L1 = 3

L1 distance

L2Unexpanded = 4

evaluate as dist_ij += (x_ik - y-jk)^2

L2SqrtUnexpanded = 5

same as above, but inside the epilogue, perform square root operation

InnerProduct = 6

basic inner product

Linf = 7

Chebyshev (Linf) distance

Canberra = 8

Canberra distance

LpUnexpanded = 9

Generalized Minkowski distance

CorrelationExpanded = 10

Correlation distance

JaccardExpanded = 11

Jaccard distance

HellingerExpanded = 12

Hellinger distance

Haversine = 13

Haversine distance

BrayCurtis = 14

Bray-Curtis distance

JensenShannon = 15

Jensen-Shannon distance

HammingUnexpanded = 16

Hamming distance

KLDivergence = 17

KLDivergence

RusselRaoExpanded = 18

RusselRao

DiceExpanded = 19

Dice-Sorensen distance

BitwiseHamming = 20

Bitstring Hamming distance

Precomputed = 100

Precomputed (special value)